Struct ssz_rs::prelude::Bitvector

pub struct Bitvector<const N: usize>(/* private fields */);

A homogenous collection of a fixed number of boolean values.

NOTE: a Bitvector of length 0 is illegal.

NOTE: once const_generics and const_evaluatable_checked features stabilize, this type can use something like bitvec::array::BitArray<T, {N / 8}> where T: BitRegister, [T; {N / 8}]: BitViewSized

Refer: https://stackoverflow.com/a/65462213

Implementations

impl<const N: usize> Bitvector<N>

pub fn get(&mut self, index: usize) -> Option<bool>

Return the bit at index. None if index is out-of-bounds.

pub fn set(&mut self, index: usize, value: bool) -> Option<bool>

Set the bit at index to value. Return the previous value or None if index is out-of-bounds.